Adding `metadata` management tools (#21)
We want to leverage the LLMs to infer and construct KumoRFM metadata,
including select columns, setup stypes, and modify links:
- `inspect_metadata`: returns full metadata spec to the LLM
- `update_metadata`: updates metadata given `GraphMetadata` model. In
the current implementation `update_metadata` throws away the old
LocalGraph and reconstructs all objects. In the future we can add a
`merge` mode which updates only part of the metadata.

---------

Co-authored-by: Matthias Fey <matthias.fey@tu-dortmund.de>